Fixed : invalid color with some PGS subtitles
git-svn-id: https://mpc-hc.svn.sourceforge.net/svnroot/mpc-hc/trunk@1039 10f7b99b-c216-0410-bff0-8a66a9350fd8

+++ src/subtitles/HdmvSub.h
@@ -115,7 +115,7 @@ public:
 		int			m_nRLEDataSize;
 		int			m_nRLEPos;
 		int			m_nColorNumber;
-		long		m_Colors[256];
+		DWORD		m_Colors[256];
 	};
 
 	CHdmvSub();
